Move a couple of relationship and declared_attr definitions from sqlite_db to the mixin, so they are available for both SQLite and Postgres
